Moved OAuthClient to its own module
